+++ sdk/communication/Azure.Communication.Administration/CHANGELOG.md
@@ -6,8 +6,11 @@
 - Added support to create CommunicationIdentityClient with TokenCredential.
 - Added support to create CommunicationIdentityClient with AzureKeyCredential.
 - Added support to create PhoneNumberAdministrationClient with AzureKeyCredential.
+- Added support to create PhoneNumberAdministrationClient with TokenCredential
+
 ### Fixed
 - Issue with paging results not pulling next pages
+
 ### Breaking
 - CommunicationTokenScope.Pstn is removed.